Add check whether `core/include/Realm` exists before copy

+++ build.sh
@@ -691,7 +691,9 @@ case "$COMMAND" in
           # CocoaPods doesn't support multiple header_mappings_dir, so combine
           # both sets of headers into a single directory
           rm -rf include
-          cp -R core/include/realm core/include/Realm
+          if [ ! -e core/include/Realm ]; then
+            cp -R core/include/realm core/include/Realm
+          fi
           cp -R core/include include
           mkdir -p include/Realm
           cp Realm/*.{h,hpp} include/Realm
